// apps/api/src/modules/whatsapp/entities/whatsapp-template.entity.d.ts
export declare class WhatsappTemplate {
    id: string;
    name: string;
    category: string;
    language: string;
    templateId: string | null;
    status: string;
    headerType: string | null;
    headerValue: string | null;
    body: string;
    footer: string | null;
    buttons: object | null;
    variables: object | null;
    rejectionReason: string | null;
    createdBy: string | null;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=whatsapp-template.entity.d.ts.map