// apps/api/src/modules/whatsapp/entities/whatsapp-campaign.entity.d.ts
export declare class WhatsappCampaign {
    id: string;
    name: string;
    description: string | null;
    status: string;
    type: string;
    templateId: string | null;
    segmentFilter: object | null;
    scheduledAt: Date | null;
    sentAt: Date | null;
    completedAt: Date | null;
    totalRecipients: number;
    sentCount: number;
    deliveredCount: number;
    readCount: number;
    failedCount: number;
    clickCount: number;
    conversionCount: number;
    metadata: object | null;
    createdBy: string | null;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=whatsapp-campaign.entity.d.ts.map